pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wechat-ai-daily"
version = "2.0.0beta"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.15.0",
    "pyautogui>=0.9.54",
    "pyperclip>=1.8.2",
    "pynput>=1.8.1",
    "requests>=2.31.0",
    "pillow>=12.1.0",
    "opencv-python>=4.8.0",
    "pygetwindow>=0.0.9",
    "rich>=13.7.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
    "psutil>=7.2.1",
    "pydantic>=2.0",
    "beautifulsoup4>=4.12",
    "PyQt6>=6.6.0",
    "ruamel-yaml>=0.18.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.18.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=1.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/wechat_ai_daily"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
# 实时显示 logging 输出
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s - %(levelname)s - %(message)s"
